refactor: extract responsible gaming assessment functions to separate file and simplify service implementation
Move assessSelfExclusion/assessDepositLimits/assessBettingLimits/assessTimeLimits/assessCoolingOffPeriods/assessInterventions/assessEducation/assessRGViolations functions from service.go to assessments.go. Replace method calls with standalone function calls in PerformResponsibleGamingAssessment. Remove EventBus/SeverityLevel/FindingStatus/ResponsibleGaming/SelfExclusionRecord/DepositLimit
