+++ internal/security/responsiblegaming/types.go
@@ -1,0 +1,239 @@
+package security
+
+import (
+	"context"
+	"time"
+
+	"github.com/shopspring/decimal"
+)
+
+// EventBus interface for publishing events
+type EventBus interface {
+	Publish(topic string, data interface{}) error
+}
+
+// ResponsibleGamingService handles responsible gaming operations
+type ResponsibleGamingService struct {
+	eventBus EventBus
+}
+
+// NewResponsibleGamingService creates a new responsible gaming service
+func NewResponsibleGamingService(eventBus EventBus) *ResponsibleGamingService {
+	return &ResponsibleGamingService{
+		eventBus: eventBus,
+	}
+}
+
+// assessSelfExclusion performs self-exclusion assessment
+func (s *ResponsibleGamingService) assessSelfExclusion(ctx context.Context) []SelfExclusionRecord {
+	// Implementation stub
+	return []SelfExclusionRecord{}
+}
+
+// assessDepositLimits performs deposit limits assessment
+func (s *ResponsibleGamingService) assessDepositLimits(ctx context.Context) []DepositLimit {
+	// Implementation stub
+	return []DepositLimit{}
+}
+
+// assessBettingLimits performs betting limits assessment
+func (s *ResponsibleGamingService) assessBettingLimits(ctx context.Context) []BettingLimit {
+	// Implementation stub
+	return []BettingLimit{}
+}
+
+// assessTimeLimits performs time limits assessment
+func (s *ResponsibleGamingService) assessTimeLimits(ctx context.Context) []TimeLimit {
+	// Implementation stub
+	return []TimeLimit{}
+}
+
+// assessCoolingOffPeriods performs cooling off periods assessment
+func (s *ResponsibleGamingService) assessCoolingOffPeriods(ctx context.Context) []CoolingOffRecord {
+	// Implementation stub
+	return []CoolingOffRecord{}
+}
+
+// assessInterventions performs interventions assessment
+func (s *ResponsibleGamingService) assessInterventions(ctx context.Context) []Intervention {
+	// Implementation stub
+	return []Intervention{}
+}
+
+// assessEducation performs education assessment
+func (s *ResponsibleGamingService) assessEducation(ctx context.Context) []EducationMaterial {
+	// Implementation stub
+	return []EducationMaterial{}
+}
+
+// assessRGViolations performs responsible gaming violations assessment
+func (s *ResponsibleGamingService) assessRGViolations(ctx context.Context) []RGViolation {
+	// Implementation stub
+	return []RGViolation{}
+}
+
+// SeverityLevel represents the severity of a security finding
+type SeverityLevel string
+
+const (
+	SeverityCritical SeverityLevel = "CRITICAL"
+	SeverityHigh     SeverityLevel = "HIGH"
+	SeverityMedium   SeverityLevel = "MEDIUM"
+	SeverityLow      SeverityLevel = "LOW"
+	SeverityInfo     SeverityLevel = "INFO"
+)
+
+// FindingStatus represents the status of a security finding
+type FindingStatus string
+
+const (
+	FindingStatusOpen       FindingStatus = "OPEN"
+	FindingStatusInProgress FindingStatus = "IN_PROGRESS"
+	FindingStatusResolved   FindingStatus = "RESOLVED"
+	FindingStatusAccepted   FindingStatus = "ACCEPTED"
+)
+
+// ResponsibleGaming represents responsible gaming compliance
+type ResponsibleGaming struct {
+	ID                string                `json:"id"`
+	ComplianceScore   float64               `json:"compliance_score"`
+	LastAssessment    time.Time             `json:"last_assessment"`
+	NextAssessment    time.Time             `json:"next_assessment"`
+	SelfExclusion     []SelfExclusionRecord `json:"self_exclusion"`
+	DepositLimits     []DepositLimit        `json:"deposit_limits"`
+	BettingLimits     []BettingLimit        `json:"betting_limits"`
+	TimeLimits        []TimeLimit           `json:"time_limits"`
+	CoolingOffPeriods []CoolingOffRecord    `json:"cooling_off_periods"`
+	Interventions     []Intervention        `json:"interventions"`
+	Education         []EducationMaterial   `json:"education"`
+	Violations        []RGViolation         `json:"violations"`
+	Recommendations   []string              `json:"recommendations"`
+}
+
+// SelfExclusionRecord represents a self-exclusion record
+type SelfExclusionRecord struct {
+	ID        string    `json:"id"`
+	UserID    string    `json:"user_id"`
+	Duration  string    `json:"duration"`
+	StartDate time.Time `json:"start_date"`
+	EndDate   time.Time `json:"end_date"`
+	Status    string    `json:"status"`
+	Reason    string    `json:"reason"`
+}
+
+// DepositLimit represents a deposit limit
+type DepositLimit struct {
+	ID         string          `json:"id"`
+	UserID     string          `json:"user_id"`
+	Type       string          `json:"type"`
+	Amount     decimal.Decimal `json:"amount"`
+	Period     string          `json:"period"`
+	Status     string          `json:"status"`
+	CreatedAt  time.Time       `json:"created_at"`
+	ModifiedAt time.Time       `json:"modified_at"`
+}
+
+// BettingLimit represents a betting limit
+type BettingLimit struct {
+	ID         string          `json:"id"`
+	UserID     string          `json:"user_id"`
+	Type       string          `json:"type"`
+	Amount     decimal.Decimal `json:"amount"`
+	Period     string          `json:"period"`
+	Status     string          `json:"status"`
+	CreatedAt  time.Time       `json:"created_at"`
+	ModifiedAt time.Time       `json:"modified_at"`
+}
+
+// TimeLimit represents a time limit
+type TimeLimit struct {
+	ID         string        `json:"id"`
+	UserID     string        `json:"user_id"`
+	Type       string        `json:"type"`
+	Duration   time.Duration `json:"duration"`
+	Period     string        `json:"period"`
+	Status     string        `json:"status"`
+	CreatedAt  time.Time     `json:"created_at"`
+	ModifiedAt time.Time     `json:"modified_at"`
+}
+
+// CoolingOffRecord represents a cooling off period record
+type CoolingOffRecord struct {
+	ID        string        `json:"id"`
+	UserID    string        `json:"user_id"`
+	Duration  time.Duration `json:"duration"`
+	StartDate time.Time     `json:"start_date"`
+	EndDate   time.Time     `json:"end_date"`
+	Status    string        `json:"status"`
+	Reason    string        `json:"reason"`
+}
+
+// Intervention represents an intervention record
+type Intervention struct {
+	ID      string    `json:"id"`
+	UserID  string    `json:"user_id"`
+	Type    string    `json:"type"`
+	Trigger string    `json:"trigger"`
+	Action  string    `json:"action"`
+	Outcome string    `json:"outcome"`
+	Date    time.Time `json:"date"`
+	Agent   string    `json:"agent"`
+}
+
+// EducationMaterial represents educational material
+type EducationMaterial struct {
+	ID        string    `json:"id"`
+	Title     string    `json:"title"`
+	Type      string    `json:"type"`
+	Content   string    `json:"content"`
+	Language  string    `json:"language"`
+	Views     int       `json:"views"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
+
+// RGViolation represents a responsible gaming violation
+type RGViolation struct {
+	ID          string        `json:"id"`
+	Type        string        `json:"type"`
+	Description string        `json:"description"`
+	Severity    SeverityLevel `json:"severity"`
+	UserID      string        `json:"user_id"`
+	Date        time.Time     `json:"date"`
+	Status      FindingStatus `json:"status"`
+	Resolved    time.Time     `json:"resolved,omitempty"`
+}
+
+// ResponsibleGamingConfig represents responsible gaming configuration
+type ResponsibleGamingConfig struct {
+	MaxDailyStake         decimal.Decimal `json:"max_daily_stake"`
+	MaxWeeklyStake        decimal.Decimal `json:"max_weekly_stake"`
+	MaxMonthlyStake       decimal.Decimal `json:"max_monthly_stake"`
+	MaxSessionDuration    time.Duration   `json:"max_session_duration"`
+	MinCoolingOffDuration time.Duration   `json:"min_cooling_off_duration"`
+	AutoIntervention      bool            `json:"auto_intervention"`
+	RequiredBreaks        bool            `json:"required_breaks"`
+	BreakInterval         time.Duration   `json:"break_interval"`
+	MinAge                int             `json:"min_age"`
+}
+
+// ProcessRequest processes a responsible gaming request
+func (s *ResponsibleGamingService) ProcessRequest(ctx context.Context, request interface{}) error {
+	// Implementation stub
+	return nil
+}
+
+// GetUserGamingProfile retrieves a user's gaming profile
+func (s *ResponsibleGamingService) GetUserGamingProfile(ctx context.Context, userID string) (interface{}, error) {
+	// Implementation stub
+	return map[string]interface{}{
+		"user_id": userID,
+		"profile": map[string]interface{}{},
+	}, nil
+}
+
+// SetGamingLimits sets gaming limits for a user
+func (s *ResponsibleGamingService) SetGamingLimits(ctx context.Context, userID string, limits interface{}) error {
+	// Implementation stub
+	return nil
+}
